Added ReadMe.txt
	I set out to create a game that incorporates my prompt, "Reliable
WiFi". Also, I set out to create some action and I did this in a form
of attacking. Ultimately, the game accomplishes what I set out to do.

	I had many challenges and problems where I got stuck during the
creation of my Final Inform Project. For instance, I had a problem with
having scenery (the WiFi) in two different rooms (McDonalds and
Starbucks). First, I went on the wiki to learn how to do this, but when
I incorporated it I had problems. It took a great deal of time and help
(Mr. and Ms. Kiang) to fix this.

	My 1st puzzle is a door and key puzzle. Basically, the player has to
examine the desk, and open the drawer to find the key that will open
the door connecting to the Boss's Office. The 2nd puzzle I have in my
game is a darkness puzzle. Basically, the player has to find the
flashlight, and know where (in the Electric Room) in order to succeed.
My 3rd and final puzzle is my NPC (Hal). In other words, the player has
to find the gun, talk to the NPC in order to kill the NPC.

	The gameplay mechanics in my game are apparent. For example, I have
navigation, failure, and accomplishment dynamics in my game. The
navigation being the darkness in the Electric Room. Secondly, the
failure being the NPC if you decide to attack without the gun.

	During the day we had the Beta tests. I had a conflict so I was not
able to have one of my classmates test my game. So I had one of
friend's test my game. The feedback affected my game in a positive way
because I was able to get multiple perspectives on the way my game was.
Not only this, but I was able to do some minor fixes, as well as touch
up on the descriptions.

	I was able to give Cole Hulick feedback on his game. For example, I
gave him feedback on the way his puzzles worked in his game. Not only
this, but also what he should name his objects because in his game it
was difficult to make things work. For instance, there was a bridge
that you had to say exactly cross bridge in order to cross it. You
could not say walk on bridge or walk over bridge. These are only a
couple things I helped Cole with.

	I am most proud of the descriptions in my game. In fact, in both my
games, the House and this one, I am proud of my descriptions because
they were really well thought-out.

	Next time, I would probably try to create a game with even more
difficult puzzles. This is because in my game I have a couple easy
puzzles and only a few difficult ones. If I had the chance to do this
again, I feel that a few more difficult puzzles would make it more
challenging and interesting.